+++ src/chanjo2/crud/cases.py
@@ -1,9 +1,13 @@
-from typing import List
+from typing import List, Union
 
+from sqlalchemy import delete
+from sqlalchemy.engine.cursor import CursorResult
 from sqlalchemy.orm import Session, query
 
 from chanjo2.models.pydantic_models import Case, CaseCreate
 from chanjo2.models.sql_models import Case as SQLCase
+from chanjo2.models.sql_models import CaseSample
+from chanjo2.models.sql_models import Sample as SQLSample
 
 
 def filter_cases_by_name(cases: query.Query, case_name: str) -> SQLCase:
@@ -18,7 +22,6 @@ def get_cases(db: Session, limit: int = 100) -> List[SQLCase]:
 
 def get_case(db: Session, case_name: str) -> Case:
     """Return a specific case by providing its name."""
-
     pipeline = {"filter_cases": filter_cases_by_name}
     query = db.query(SQLCase)
 
@@ -38,9 +41,38 @@ def count_cases(db: Session) -> int:
     return db.query(SQLCase).count()
 
 
+def delete_entry(db: Session, table: Union[SQLCase, SQLSample], id: int) -> int:
+    """Deletes a Sample or Case entry from the database."""
+    delete_stmt: Delete = delete(table).where(table.id == id)
+    result: CursorResult = db.execute(delete_stmt)
+    db.commit()
+    return result.rowcount
+
+
+def delete_case_sample_ref(db: Session, case_id: int, sample_id: int) -> int:
+    """Delete the association of a Sample to a Case in the CaseSample table."""
+    delete_stmt: Delete = delete(CaseSample).where(
+        CaseSample.c.case_id == case_id, CaseSample.c.sample_id == sample_id
+    )
+    result: CursorResult = db.execute(delete_stmt)
+    db.commit()
+    return result.rowcount
+
+
 def delete_case(db: Session, case_name: str) -> int:
     """Delete a case with the supplied name."""
-    nr_cases_before_deletion = count_cases(db=db)
-    db.delete(get_case(db=db, case_name=case_name))
-    db.commit()
-    return nr_cases_before_deletion - count_cases(db=db)
+    db_case: SQLCase = db.query(SQLCase).where(SQLCase.name == case_name).first()
+
+    # Delete samples linked uniquely to this case
+    for db_sample in db_case.samples:
+        nr_linked_cases: int = (
+            db.query(CaseSample).where(CaseSample.c.sample_id == db_sample.id).count()
+        )
+        # Remove case-sample association in CaseSample table:
+        delete_case_sample_ref(db=db, case_id=db_case.id, sample_id=db_sample.id)
+
+        if nr_linked_cases == 1:  # Sample linked only to this case
+            delete_entry(db=db, table=SQLSample, id=db_sample.id)
+
+    # Delete case
+    return delete_entry(db=db, table=SQLCase, id=db_case.id)