+++ src/chanjo2/models/sql_models.py
@@ -1,37 +1,39 @@
-from sqlalchemy import Column, DateTime, Enum, ForeignKey, Integer, String, Index
+from sqlalchemy import Column, DateTime, Enum, ForeignKey, Integer, String, Index, Table
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 
 from chanjo2.dbutil import Base
 from chanjo2.models.pydantic_models import Builds
 
+CaseSample = Table(
+    "case_sample",
+    Base.metadata,
+    Column("case_id", Integer, ForeignKey("cases.id"), primary_key=True),
+    Column("sample_id", Integer, ForeignKey("samples.id"), primary_key=True),
+)
+
 
 class Case(Base):
-    """Used to define a group of samples."""
+    """Used to define a case containing samples."""
 
     __tablename__ = "cases"
-
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String(64), nullable=False, unique=True, index=True)
     display_name = Column(String(64), nullable=True, unique=False)
-
-    samples = relationship("Sample", cascade="all,delete", backref="case")
+    samples = relationship("Sample", secondary=CaseSample, back_populates="cases")
 
 
 class Sample(Base):
-    """Used to define a single sample belonging to a Case"""
+    """Used to define a single sample belonging to a Case."""
 
     __tablename__ = "samples"
-
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String(64), nullable=False, unique=True, index=True)
     display_name = Column(String(64), nullable=True, unique=False)
     track_name = Column(String(64), nullable=True, unique=False)
-    case_id = Column(
-        Integer, ForeignKey("cases.id", ondelete="CASCADE"), nullable=False
-    )
     coverage_file_path = Column(String(512), nullable=False)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
+    cases = relationship("Case", secondary=CaseSample, back_populates="samples")
 
 
 class Interval(Base):